Fix proptype error
