+++ src/reducers/index.js
@@ -1,10 +1,8 @@
-import { combineReducers, createStore } from 'redux';
+import { combineReducers } from 'redux';
 import booksReducer from './books';
 
-const configureStore = () => createStore(
-  combineReducers({
-    books: booksReducer,
-  }),
-);
+const configureStore = combineReducers({
+  books: booksReducer,
+});
 
 export default configureStore;
